refactor: extract ScoreAndRank (Stage 3-4) into src/domain/usecases
Fifth and largest extraction in the hexagon-cleanup plan. Pulls the
rest of the scoring/ranking logic out of stagedSearch.ts:

- promptFitBreakdown, inferRepoArtifactType — Stage 3 prompt-fit scoring
- readmeQualityScore, freshnessScore, dependencyHealthFor,
  healthBreakdown, repoHealthScore, decayLabelFor — health/freshness/
  decay/dependency signals
- rankingWeights, confidenceLabel — Stage 4 composite ranking
- persistIntelligence — the RepoIntelligencePort snapshot/health-score
  writes, moved with the scoring logic that produces the data it persists

ScoreAndRank.execute(candidates, classification, intent) runs prompt-fit
filtering, composite scoring (dropping anything below the health floor
or decayed to Abandoned), sorts by final score, persists intelligence
data for survivors, and returns both the ranked list and the pre-filter
promptFitPassedCount (stageCounts needs both the Stage 3 and Stage 4
counts, same reason DiscoverRepos returns stage1Raw alongside enriched).

stagedSearch.ts drops from 787 lines (after the DiscoverRepos/
ApplyQualityGates extractions) to 265 — now just classifyIntent (Stage 0
classification, staying put for now), runStagedSearch (composition:
wires DiscoverRepos -> applyQualityGates -> ScoreAndRank -> Stage 5
trim/confidence-label), and renderStagedSearch (CLI presentation only).
RankedRepo's consumer in agent.ts updated to import from the new
location instead of re-exported through stagedSearch.ts.

17 new tests (3 for confidenceLabel's threshold boundaries, 6 for
ScoreAndRank.execute — a strong candidate ranking + persisting
correctly, prompt-fit rejection, health-floor rejection, Abandoned-decay
rejection even with a passing health score, sort order across two
candidates, and a persistence-failure case that must not lose results).
One fixture needed a second pass: dependencyHealthFor defaults to
"Clean" (worth 20 of the health score's 100 points) whenever there's no
red-flag text and no rootContents, so a "weak repo" fixture with an
empty description didn't actually clear the floor until deprecation
language and dependency signals were added deliberately. Full suite:
65/65 green, tsc --noEmit clean.
